1. Update color of UI 2. Fix Event page

+++ app/src/main/java/my/com/taruc/fitnesscompanion/UI/EventPage.java
@@ -2,7 +2,9 @@ package my.com.taruc.fitnesscompanion.UI;
 
 import android.app.ProgressDialog;
 import android.content.Context;
+import android.os.AsyncTask;
 import android.os.Bundle;
+import android.os.Handler;
 import android.support.v7.app.ActionBarActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -20,6 +22,7 @@ import butterknife.Bind;
 import butterknife.ButterKnife;
 import my.com.taruc.fitnesscompanion.Adapter.EventAdapter;
 import my.com.taruc.fitnesscompanion.Classes.Event;
+import my.com.taruc.fitnesscompanion.Classes.TaskCanceler;
 import my.com.taruc.fitnesscompanion.ConnectionDetector;
 import my.com.taruc.fitnesscompanion.Database.EventDA;
 import my.com.taruc.fitnesscompanion.R;
@@ -38,15 +41,14 @@ public class EventPage extends ActionBarActivity {
     private EventAdapter eventAdapter;
     private EventDA eventDA;
     private RetrieveRequest mRetrieveRequest;
-    private ArrayList<Event> eventArrayList;
+    private ArrayList<Event> eventArrayList = new ArrayList<>();
     private ArrayList<Event> eventArrayListFromServer;
 
     private ConnectionDetector mConnectionDetector;
     private ShowAlert alert = new ShowAlert();
-    private ProgressDialog mProgressDialog;
 
     private Context context;
-    private Timer timer = new Timer();
+    private ProgressDialog mProgressDialog;
     
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -56,42 +58,38 @@ public class EventPage extends ActionBarActivity {
         ButterKnife.bind(this);
         textViewTitle.setText(R.string.eventTitle);
 
-        //do server request for event
         mConnectionDetector = new ConnectionDetector(this);
         mRetrieveRequest = new RetrieveRequest(this);
         eventDA = new EventDA(this);
+
         eventArrayList = eventDA.getAllEvent();
-        eventArrayListFromServer = mRetrieveRequest.fetchAllEventInBackground();
 
         if(!mConnectionDetector.isConnectingToInternet()){
             alert.showAlertDialog(this, "Internet Error", "No Internet", false);
         } else {
-            if (eventArrayList.isEmpty()) {
-                mProgressDialog = ProgressDialog.show(this, "Synchronizing", "Sync with server....Please Wait.", true);
-                timer.schedule(new TimerTask() {
-                    @Override
-                    public void run() {
-                        runOnUiThread(new Runnable() {
-                            @Override
-                            public void run() {
-                                eventDA.addEventArrayList(eventArrayListFromServer);
-                                mProgressDialog.dismiss();
-                                eventAdapter = new EventAdapter(context, eventArrayListFromServer);
-                                try {
-                                    mRecyclerView.swapAdapter(eventAdapter, true);
-                                    timer.cancel();
-                                } catch (Exception ex) {
-                                    Log.i("EventErr", ex.getMessage());
-                                }
-                            }
-                        });
-                    }
-                }, 500, 10);
+            mProgressDialog = ProgressDialog.show(this, "Synchronizing", "Sync with server....Please Wait.", true);
+            new Thread(new Runnable() {
+                @Override
+                public void run() {
 
-            } else if (eventArrayList.size() != eventArrayListFromServer.size()) {
-                eventDA.deleteAll();
-                eventDA.addEventArrayList(eventArrayListFromServer);
-            }
+                    eventArrayListFromServer = mRetrieveRequest.fetchAllEventInBackground();
+                    if (eventArrayList.size() != eventArrayListFromServer.size() && !eventArrayListFromServer.isEmpty()) {
+                        try{
+                            Thread.sleep(1000);
+                        }catch (Exception ex){
+                            Log.i("Event","Thread Exception");
+                        }
+                        eventDA.deleteAll();
+                        eventDA.addEventArrayList(eventArrayListFromServer);
+                        eventArrayList = eventDA.getAllEvent();
+
+                        eventAdapter = new EventAdapter(context, eventArrayList);
+                        mRecyclerView.swapAdapter(eventAdapter, true);
+                    }
+                    mProgressDialog.dismiss();
+                }
+
+            }).start();
         }
 
         mRecyclerView.setHasFixedSize(true);
@@ -104,4 +102,8 @@ public class EventPage extends ActionBarActivity {
         this.finish();
     }
 
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+    }
 }
